Build fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@65742 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DrvAudio.cpp
@@ -1583,7 +1583,7 @@ static int drvAudioStreamCaptureNonInterleaved(PDRVAUDIO pThis, PPDMAUDIOSTREAM 
         }
 
         if (cbFree > cbBuf) /* Limit to buffer size. */
-            cbFree = cbBuf;
+            cbFree = (uint32_t)cbBuf;
 
         uint32_t cbCaptured;
         rc = pThis->pHostDrvAudio->pfnStreamCapture(pThis->pHostDrvAudio, pHstStream->pvBackend,
@@ -1597,7 +1597,7 @@ static int drvAudioStreamCaptureNonInterleaved(PDRVAUDIO pThis, PPDMAUDIOSTREAM 
         {
             Assert(cbCaptured <= cbBuf);
             if (cbCaptured > cbBuf) /* Paranoia. */
-                cbCaptured = cbBuf;
+                cbCaptured = (uint32_t)cbBuf;
 
             uint32_t csCaptured = 0;
             rc = AudioMixBufWriteCirc(&pHstStream->MixBuf, auBuf, cbCaptured, &csCaptured);
